+++ lib/rspec_tracer/rspec_reporter.rb
@@ -6,7 +6,7 @@ module RSpecTracer
       RSpecTracer.coverage_reporter.record_coverage
       RSpecTracer.start_example_trace
 
-      super(example)
+      super
     end
 
     def example_finished(example)
@@ -14,28 +14,28 @@ module RSpecTracer
       RSpecTracer.stop_example_trace(example_id)
       RSpecTracer.coverage_reporter.compute_diff(example_id)
 
-      super(example)
+      super
     end
 
     def example_passed(example)
       example_id = example.metadata[:rspec_tracer_example_id]
       RSpecTracer.runner.on_example_passed(example_id, example.execution_result)
 
-      super(example)
+      super
     end
 
     def example_failed(example)
       example_id = example.metadata[:rspec_tracer_example_id]
       RSpecTracer.runner.on_example_failed(example_id, example.execution_result)
 
-      super(example)
+      super
     end
 
     def example_pending(example)
       example_id = example.metadata[:rspec_tracer_example_id]
       RSpecTracer.runner.on_example_pending(example_id, example.execution_result)
 
-      super(example)
+      super
     end
   end
 end